+++ tests/testthat/test-modifiers.R
@@ -8,3 +8,19 @@ test_that("mod_infinitable", {
   is_incrementing_inf <- mod_infinitable(is_incrementing)
   expect_true(is_incrementing_inf(x))
 })
+
+test_that("mod_nullable works with custom function", {
+  sch <- schema(
+    c(a, b) ~ mod_nullable(~all(.x %in% c("a", "b", "c")))
+  )
+
+  expect_no_error({
+    check_schema(
+      data.frame(
+        a = c(NA, "a", "c"),
+        b = c("a", "b", "c")
+      ),
+      sch
+    )
+  })
+})