+++ NEWS.md
@@ -1,3 +1,13 @@
+# schematic 0.1.1
+
+### Bug fixes
+
+- Fixed issue where lots of column names in the message were being replaced with '...'.
+
+- Fixed issue where checking on a column named `e` triggered error in `check_schema()`.
+
+- Fixed issue where modifiers wouldn't work with custom predicates.
+
 # schematic 0.1.0
 
 ### CRAN release